+++ playground/src/data/sections.ts
@@ -1141,7 +1141,7 @@ export const propkitSections: Section[] = [
         name: "Default",
         markup: `<div class="prop-panel">
   <fig-button data-playground-ignore-controls="true" onclick="const d=this.nextElementSibling; d.hasAttribute('modal') ? d.showModal() : d.show();">Open Plugin</fig-button>
-  <dialog is="fig-dialog" title="Plugin" open handle="fig-header" position="center center" data-playground-hide-field style="width: 240px;">
+  <dialog is="fig-dialog" title="Plugin" open handle="fig-header" position="center center" autoresize data-playground-hide-field style="width: 240px;">
     <iframe src="/propkit/iframe.html"></iframe>
   </dialog>
 </div>`,